Image details screen - not complete yet - please see below:
Hey Ryan - here's where I am with the image details screen - please see app.tsx and image-details.tsx.  I think the dynamic routing is working in app.tsx.  In image-details.tsx, I've got a query to return some info about an image based on its ID - this is the "RetrieveImage" query.  However, you'll also see a query called "somethingelse" - I just tried a query for a tenant instead of for an image since there aren't currently any images to query.  But I think it's the same idea.

The issue I'm running into is that I need a way to wait to print the info about the image (returned by the query) until we know it's ready - that is, once the "loading" boolean is false.  The Apollo documentation has an example of this using an if-statement that returns null if the page is still loading - please see line 48 in image-details.tsx.  But I'm not able to use the "return null" statement without getting a type error saying that this couldn't be assigned to type RouteComponentProps.  Do you have any ideas?  Thanks.